created basic flashcard functionality with dummy data
