+++ client/src/App.jsx
@@ -1,16 +1,14 @@
 import { createBrowserRouter, createRoutesFromElements, Route, RouterProvider } from "react-router-dom"
-import { useState } from 'react'
-
 // Pages and Components Required for Application 
 import Home from './pages/Home'
 import MyDesk from './pages/MyDesk'
 import InterviewPrep from './components/InterviewPrep'
 import Login from './components/Login'
 import Signup from './components/Signup'
-import FlashcardList from "./components/FlashcardList"
 import Flashcard from "./components/Flashcard"
 import Navbar from './components/Navbar'
 import JobList from './components/JobList'
+import './App.css';
 // Routes
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -18,7 +16,7 @@ const router = createBrowserRouter(
       <Route path='/' element={<Home />} />
       <Route path='mydesk' element={<MyDesk />} />
       <Route path='interviewprep' element={<InterviewPrep />} />
-      <Route path='flashcards' element={<FlashcardList />} />
+      <Route path='flashcards' element={<InterviewPrep />} />
       <Route path='flashcard' element={<Flashcard />} />
       <Route path='login' element={<Login />} />
       <Route path='signup' element={<Signup />} />
@@ -28,38 +26,13 @@ const router = createBrowserRouter(
 )
 
 function App() {
-  const [flashcards] = useState(SAMPLE_FLASHCARDS)
   return (
     <div>
     <Navbar />
     <RouterProvider router={router} />
-    <FlashcardList flashcards={flashcards}/>
+    {/* <FlashcardList flashcards={flashcards}/> */}
     </div>
   )
 }
 
-const SAMPLE_FLASHCARDS = [
-  {
-    id: 1,
-    question: 'Sample question',
-    answer: 'Sample answer',
-    options: [
-      'A',
-      'B',
-      'C',
-      'D'
-    ]
-  },
-  {
-  id: 2,
-  question: 'Question 2',
-  answer: 'Answer 2',
-  options: [
-    '1',
-    '2',
-    '3',
-    '4'
-    ]
-  },
-]
 export default App
