+++ client/src/components/FlashcardList.jsx
@@ -1,10 +1,11 @@
-// import Flashcard from "./Flashcard"
+import Flashcard from './Flashcard'
 
-// const { flashcards } = Flashcard;
-
-export default function FlashcardList() {
+export default function FlashcardList({ flashcards }) {
     return (
         <div className="card-grid">
+            {flashcards.map((flashcard => {
+                return <Flashcard flashcard={flashcard} key={flashcard.id} />
+            }))}
         </div>
     )
 }
